+++ src/main/scala/com/atomist/rug/ts/TypeGenerator.scala
@@ -1,53 +1,107 @@
 package com.atomist.rug.ts
 
-import java.util.Objects
+import java.io.PrintWriter
 
 import com.atomist.param.SimpleParameterValues
-import com.atomist.rug.spi.{SimpleTypeRegistry, TypeOperation, Typed}
+import com.atomist.rug.spi.{SimpleTypeRegistry, TypeOperation, TypeRegistry, Typed}
 import com.atomist.source.ArtifactSource
-import com.atomist.util.lang.JavaHelpers
+import com.atomist.util.Utils
+import com.atomist.util.lang.JavaHelpers._
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.scala.DefaultScalaModule
 import com.fasterxml.jackson.module.scala.experimental.ScalaObjectMapper
+import org.apache.commons.io.IOUtils
+
+object TypeGenerator {
+
+  val CortexJsonLocation = "/com/atomist/rug/ts/cortex.json"
+
+  lazy val CortexJson = IOUtils.toString(getClass.getResourceAsStream(CortexJsonLocation), "UTF-8")
+
+  val DefaultCortexDir = "cortex"
+
+  val DefaultCortexStubDir = "cortex/stub"
+
+  lazy val ExtendedTypes: TypeRegistry = {
+    val types = new TypeGenerator(DefaultCortexDir, DefaultCortexDir).extract(CortexJson)
+    new SimpleTypeRegistry(types)
+  }
+
+}
+
+object TypeGeneratorApp extends App {
+
+  import TypeGenerator._
+
+  val target = args.head
+  val tsig = new TypeGenerator(DefaultCortexDir, DefaultCortexStubDir)
+  val output = tsig.toNodeModule(TypeGenerator.CortexJson)
+  println(s"Generated Type module")
+  output.allFiles.foreach(f => Utils.withCloseable(new PrintWriter(target + "/" + f.path))(_.write(f.content)))
+}
 
 /**
   * Take endpoints reported in JSON from a materializer service and generate types
-  * from which we can in turn generate TypeScript interfaces.
-  * There's little error checking, as this isn't intended as production code.
   */
-class TypeGenerator(basePackage: String = "ext_model") {
+class TypeGenerator(basePackage: String, baseClassPackage: String) {
 
   private val mapper = new ObjectMapper() with ScalaObjectMapper
   mapper.registerModule(DefaultScalaModule)
 
   private def toTypeScriptFiles(json: String): ArtifactSource = {
     val types = extract(json)
-    val tig = new TypeScriptInterfaceGenerator(typeRegistry =
-      new SimpleTypeRegistry(types))
-    tig.generate("types", SimpleParameterValues("output_path", "Types.ts"))
+    val typeRegistry = new SimpleTypeRegistry(types)
+    //types.foreach(println(_))
+    val tig = new TypeScriptInterfaceGenerator(typeRegistry)
+    val tcg = new TypeScriptClassGenerator(typeRegistry)
+    tig.generate("types", SimpleParameterValues("output_path", "Types.ts")).withPathAbove(basePackage) +
+      tcg.generate("types", SimpleParameterValues("output_path", "Types.ts")).withPathAbove(baseClassPackage)
   }
 
   /**
     * Return a valid node module with these files in it, under base package
     */
   def toNodeModule(json: String): ArtifactSource = {
-    toTypeScriptFiles(json).withPathAbove(basePackage)
+    toTypeScriptFiles(json)
   }
 
   /**
     * Extract the types described in this document and output a valid node module
     */
   def extract(json: String): Set[Typed] = {
+    require(json != null)
     val doc: Map[String, List[_]] = mapper.readValue(json, classOf[Map[String, List[_]]])
 
+    def defineProperty(rawName: String, typ: String): Prop = {
+      // Do aliasing
+      val propName = toTypeScriptIdentifier(rawName)
+      val genTyp = typ match {
+        case "string" => "String"
+        case "number" => "long"
+        case "url" => "String" // TODO we should somehow use helpers for this and construct in the proxy from a string
+        case "timestamp" => "String"
+        case x => x
+      }
+      Prop(propName, genTyp)
+    }
+
+    /*
+    {"nodes":
+    {"labels":["Push"],
+    "properties":
+    [["before", "string"], ["after", "string"],
+    ["timestamp", "timestamp"], ["branch", "string"]],
+    "unique":["branch", "after"]},
+     */
     val propertyNodes: Traversable[PropertyNode] =
       doc("nodes").map(node => {
+        // A node is of form labels, properties, unique?
         val n = node.asInstanceOf[Map[String, _]]
         val props = n("properties") match {
           case l: List[List[String]]@unchecked if l.head.isInstanceOf[Seq[_]] =>
-            l.map(l => Prop(l.head, l(1)))
+            l.map(l => defineProperty(l.head, l(1)))
           case l: List[String]@unchecked =>
-            Seq(Prop(l.head, l(1)))
+            Seq(defineProperty(l.head, l(1)))
         }
 
         val pn = PropertyNode(
@@ -57,49 +111,60 @@ class TypeGenerator(basePackage: String = "ext_model") {
         pn
       })
 
-    // These are normally a - b - b
-    val relationships: Traversable[Relationship] =
-      doc("relationships") flatMap {
-        case relList: List[_]@unchecked if relList.size == 3 =>
-          val relName = toTypeScriptName(relList(1) match {
-            case s: String => s
-            case l: List[_] => Objects.toString(l.head)
-          })
-          val a = relList.head match {
-            case s: String => List(s)
-            case l: List[_] => l.map(Objects.toString)
-          }
-          val b = relList(2) match {
-            case s: String => List(s)
-            case l: List[_] => l.map(Objects.toString)
-          }
-          // TODO note we're just taking the first one of the bs: This is wrong, but otherwise
-          // We'd need to disambiguate relationship names
-          val r = Relationship(a.toSet, relName, Set(b.head))
-          Seq(r)
-        case _ => ???
-      }
+    // These are normally a - [right,cardinality] - right
+    // [["Org"], ["HAS", "1:M", ["OWNED_BY", "1:1"]], ["Repo"]]
+    val allRelationships: Traversable[Relationship] =
+    doc("relationships") flatMap {
+      case List(left: String, List(name: String, card: String), right: String) =>
+        val cardinality = Cardinality(card)
+        val relName = toTypeScriptIdentifier(name)
+        Seq(Relationship(left, relName, cardinality, right))
+      case List(left: String, List(name: String, card: String, List(backName: String, backCard: String)), right: String) =>
+        val cardinality = Cardinality(card)
+        val relName = toTypeScriptIdentifier(name)
+        val backCardinality = Cardinality(backCard)
+        val backRelName = toTypeScriptIdentifier(backName)
+        val backRel = Relationship(right, backRelName, backCardinality, left)
+        //println(s"bidirectional relationship $backRel")
+        Seq(
+          Relationship(left, relName, cardinality, right),
+          backRel
+        )
+      case x =>
+        throw new IllegalArgumentException(s"Illegal list, not able to destructure, $x")
+    }
 
-    propertyNodes.map(pn => new JsonBackedTyped(pn,
-      relationships.filter(_.as.contains(pn.labels.head)))
-    ).toSet
+    propertyNodes.map(pn => new JsonBackedTyped(pn, allRelationships)).toSet
   }
 
-  private def toTypeScriptName(name: String): String =
-    JavaHelpers.toCamelizedPropertyName(name.toLowerCase)
+  private def toTypeScriptIdentifier(name: String): String =
+    toCamelizedPropertyName(name.toLowerCase)
 }
 
 private case class Prop(name: String, typ: String)
 
 private case class PropertyNode(labels: Set[String], properties: Seq[Prop])
 
-private case class Relationship(as: Set[String], rel: String, bs: Set[String])
+private sealed trait Cardinality
 
-// TODO do we need to have multiple labels in Typed?
+private object Cardinality {
+
+  def apply(s: String): Cardinality = s match {
+    case "1:1" => OneToOne
+    case "1:M" => OneToM
+    case x => throw new IllegalArgumentException(s"Unknown cardinality: [$x]")
+  }
+}
+
+private case object OneToOne extends Cardinality
+
+private case object OneToM extends Cardinality
+
+private case class Relationship(left: String, name: String, cardinality: Cardinality, right: String)
 
 private class JsonBackedTyped(
                                properties: PropertyNode,
-                               relationships: Traversable[Relationship]
+                               allRelationships: Traversable[Relationship]
                              ) extends Typed {
 
   override val name: String = properties.labels.head
@@ -110,26 +175,35 @@ private class JsonBackedTyped(
   override def allOperations: Seq[TypeOperation] = operations
 
   override def operations: Seq[TypeOperation] = {
-    val propsOps: Seq[TypeOperation] =
-      properties.properties.map(prop => TypeOperation(
-        name = prop.name,
-        description = prop.typ,
-        readOnly = false,
-        parameters = Nil,
-        returnType = "String",
-        definedOn = null,
-        example = None
-      ))
     val relOps: Seq[TypeOperation] =
-      relationships.map(rel => TypeOperation(
-        name = rel.rel,
-        description = s"$name - ${rel.as.head} -> ${rel.bs}",
-        readOnly = false,
-        parameters = Nil,
-        returnType = rel.bs.head,
-        definedOn = null,
-        example = None
-      )).toSeq
+      allRelationships
+        .filter(_.left == name)
+        .map(rel => TypeOperation(
+          name = rel.name,
+          description = s"${rel.name} - ${rel.left} -> ${rel.right}",
+          readOnly = false,
+          parameters = Nil,
+          returnType = rel.cardinality match {
+            case OneToOne => rel.right
+            case OneToM => s"${rel.right}[]"
+          },
+          definedOn = null,
+          example = None
+        )).toSeq
+    val propsOps: Seq[TypeOperation] =
+      properties.properties
+        // Filter out properties that have the same name as a relationship, that takes precedence
+        .filterNot(p => relOps.exists(_.name == p.name))
+        .map(prop =>
+          TypeOperation(
+            name = prop.name,
+            description = prop.typ,
+            readOnly = false,
+            parameters = Nil,
+            returnType = prop.typ,
+            definedOn = null,
+            example = None
+          ))
 
     propsOps ++ relOps
   }