+++ src/main/scala/com/atomist/util/lang/TypeScriptGenerationHelper.scala
@@ -31,7 +31,7 @@ class TypeScriptGenerationHelper(indent: String = "    ")
       case "java.util.List<java.lang.String>" => "string[]"
       case "class java.lang.String" => "string"
       case "Object" => "any"
-      case "class java.lang.Object" =>  "any"
+      case "class java.lang.Object" => "any"
       case "java.util.List<com.atomist.rug.kind.core.ProjectMutableView>" => "Project[]"
       case "class com.atomist.rug.kind.core.ProjectMutableView" => "Project"
       case "java.util.List<java.lang.Object>" => "any[]"
@@ -39,12 +39,12 @@ class TypeScriptGenerationHelper(indent: String = "    ")
       case "class com.atomist.rug.kind.core.ProjectContext" => "ProjectContext"
       case "scala.collection.immutable.List<java.lang.Object>" => "any[]"
       case "java.util.List<com.atomist.rug.kind.core.FileArtifactBackedMutableView>" => "File[]"
-      case "java.util.List<com.atomist.rug.kind.java.support.PackageInfo>" => "any[]"//TODO
+      case "java.util.List<com.atomist.rug.kind.java.support.PackageInfo>" => "any[]" //TODO
       case "java.util.List<com.atomist.rug.kind.service.ServiceMutableView>" => "any[]"
       case "java.util.List<com.atomist.tree.TreeNode>" => "any[]"
       case "interface com.atomist.tree.TreeNode" => "TreeNode" // is this right?
       case "List" => "any[]" // TODO improve this
-      case "FileArtifactMutableView" => "File"   // TODO this is nasty
+      case "FileArtifactMutableView" => "File" // TODO this is nasty
       case "scala.collection.immutable.Set<java.lang.String>" => "string[]" // Nasty
       case `pathExpressionEngineClassName` => "PathExpressionEngine"
       case "class com.atomist.tree.content.text.FormatInfo" => "FormatInfo"
@@ -56,7 +56,9 @@ class TypeScriptGenerationHelper(indent: String = "    ")
         //println(s"Returning [$cname]")
         cname
       case x if tr.findByName(x).isDefined => x
-      case x => throw new UnsupportedOperationException(s"Unsupported type [$jt]. Did you export a function with this in its type signature?")
+      case x if x.endsWith("[]") && tr.findByName(x.stripSuffix("[]")).isDefined =>
+        x
+      case x => throw new UnsupportedOperationException(s"Unsupported type [$x]. Did you export a function with this in its type signature?")
     }
   }
 }