+++ src/test/scala/com/atomist/rug/rugdoc/TypeScriptClassGeneratorTest.scala
@@ -5,6 +5,7 @@ import com.atomist.rug.spi.SimpleTypeRegistry
 import com.atomist.rug.ts._
 import com.atomist.source.{ArtifactSource, FileArtifact, FileEditor}
 import org.scalatest.{FlatSpec, Matchers}
+import TypeGenerator._
 
 object TypeScriptClassGeneratorTest {
 
@@ -21,32 +22,9 @@ object TypeScriptClassGeneratorTest {
       // Note: We need to pretend we have imports that will be available
       // at runtime
       override def edit(f: FileArtifact): FileArtifact =
-      f.withContent(f.content.replace(InterfaceGenerationConfig.TestStubImports,
+      f.withContent(f.content.replace(TypeGenerationConfig.TestStubImports,
         """
-          |interface ProjectContext {}
-          |interface PathExpressionEngine {}
-          |interface TreeNode {}
-          |interface FormatInfo {}
-          |interface Addressed {
-          |   address(): string
-          |}
-          |abstract class AddressedNodeSupport implements Addressed {
-          |
-          |    private _navigatedFrom: Addressed = null
-          |    private step: string
-          |
-          |    address() {
-          |        return this._navigatedFrom ?
-          |            this._navigatedFrom.address() + this.step :
-          |            ""
-          |    }
-          |    navigatedFrom(navigatedFrom: Addressed, step: string) {
-          |        if (navigatedFrom == this)
-          |            throw new Error(`Illegal cycle at ${this}`)
-          |        this._navigatedFrom = navigatedFrom
-          |        this.step = step
-          |    }
-          |}
+          |interface GraphNode {}
         """.stripMargin))
     }
 
@@ -69,7 +47,7 @@ class TypeScriptClassGeneratorTest extends FlatSpec with Matchers {
   // We don't need to test against project types like File
   it should "generate compilable typescript classes" in {
 
-    val types = new TypeGenerator().extract(TypeGeneratorTest.CortexJson)
+    val types = new TypeGenerator(DefaultCortexDir, DefaultCortexStubDir).extract(CortexJson)
     val tr = new SimpleTypeRegistry(types)
 
     val td = new TypeScriptClassGenerator(tr)
@@ -86,8 +64,8 @@ class TypeScriptClassGeneratorTest extends FlatSpec with Matchers {
     val js = compiled.allFiles.find(_.name.endsWith(".js"))
     js shouldBe defined
 
-    val js2 = compiled.allFiles.find(_.content.contains("withResolvedBy"))
-    js2 shouldBe defined
+    //val js2 = compiled.allFiles.find(_.content.contains("withResolvedBy"))
+    //js2 shouldBe defined
     // println(js.get.content)
   }
 }